[cleanup] replace C-style casts with static_cast

+++ dune/localfunctions/lagrange/pktrace2d/pktrace2dlocalbasis.hh
@@ -42,7 +42,7 @@ namespace Dune
     PkTrace2DLocalBasis ()
     {
       for (unsigned int i=0; i<=k; i++)
-        pos[i] = (1.0*i)/std::max(k,(unsigned int)1);
+        pos[i] = (1.0*i)/std::max(k,1u);
     }
 
     //! \brief number of shape functions
